add back handleEvents types, delete old ECharts.tsx
